README update and more general cleanup.
